Move HtmlRelation back into lib/relations as it's also used for other stuff
