+++ lib/relations/Html/HtmlAlternateLink.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlAlternateLink extends HtmlRelation {
   get href() {
