+++ lib/relations/Html/HtmlAnchor.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlAnchor extends HtmlRelation {
   get href() {
