+++ lib/relations/Html/HtmlAppleTouchStartupImage.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlAppleTouchStartupImage extends HtmlRelation {
   get href() {
