+++ lib/relations/Html/HtmlApplicationManifest.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 /* Models application manifests.
  *
