+++ lib/relations/Html/HtmlAudio.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlAudio extends HtmlRelation {
   get href() {
