+++ lib/relations/Html/HtmlAuthorLink.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlAuthorLink extends HtmlRelation {
   get href() {
