+++ lib/relations/Html/HtmlCacheManifest.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlCacheManifest extends HtmlRelation {
   get href() {
