+++ lib/relations/Html/HtmlConditionalComment.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlConditionalComment extends HtmlRelation {
   constructor(config) {
