+++ lib/relations/Html/HtmlContentSecurityPolicy.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlContentSecurityPolicy extends HtmlRelation {
   inline() {
