+++ lib/relations/Html/HtmlDataBindAttribute.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlDataBindAttribute extends HtmlRelation {
   inline() {
