+++ lib/relations/Html/HtmlDnsPrefetchLink.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 /**
  * @extends {HtmlRelation}
