+++ lib/relations/Html/HtmlEdgeSideInclude.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlEdgeSideInclude extends HtmlRelation {
   get href() {
