+++ lib/relations/Html/HtmlEdgeSideIncludeSafeComment.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlEdgeSideIncludeSafeComment extends HtmlRelation {
   inline() {
