+++ lib/relations/Html/HtmlEmbed.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlEmbed extends HtmlRelation {
   get href() {
