+++ lib/relations/Html/HtmlFluidIconLink.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlFluidIconLink extends HtmlRelation {
   get href() {
