+++ lib/relations/Html/HtmlFrame.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlFrame extends HtmlRelation {
   get href() {
