+++ lib/relations/Html/HtmlIFrame.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlIFrame extends HtmlRelation {
   get href() {
