+++ lib/relations/Html/HtmlIFrameSrcDoc.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlIFrameSrcDoc extends HtmlRelation {
   inline() {
