+++ lib/relations/Html/HtmlImage.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlImage extends HtmlRelation {
   get href() {
