+++ lib/relations/Html/HtmlImageSrcSet.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlImageSrcSet extends HtmlRelation {
   inline() {
