+++ lib/relations/Html/HtmlImport.js
@@ -1,8 +1,8 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlImport extends HtmlRelation {
   static get selector() {
-    return 'link[rel~=import]';
+    return 'link[rel~=import][href]';
   }
 
   static handler(node) {
