+++ lib/relations/Html/HtmlInlineEventHandler.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlInlineEventHandler extends HtmlRelation {
   static get selector() {
