+++ lib/relations/Html/HtmlInlineFragment.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlInlineFragment extends HtmlRelation {
   inline() {
