+++ lib/relations/Html/HtmlInlineScriptTemplate.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');;
 
 class HtmlInlineScriptTemplate extends HtmlRelation {
   inline() {
