+++ lib/relations/Html/HtmlKnockoutContainerless.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlKnockoutContainerless extends HtmlRelation {
   inline() {
