+++ lib/relations/Html/HtmlLogo.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlLogo extends HtmlRelation {
   get href() {
