+++ lib/relations/Html/HtmlMetaRefresh.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlMetaRefresh extends HtmlRelation {
   get href() {
