+++ lib/relations/Html/HtmlMsApplicationConfig.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlMsApplicationConfig extends HtmlRelation {
   get href() {
