+++ lib/relations/Html/HtmlMsApplicationTileImageMeta.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlMsApplicationTileImageMeta extends HtmlRelation {
   get href() {
