+++ lib/relations/Html/HtmlNoscript.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlNoscript extends HtmlRelation {
   inline() {
