+++ lib/relations/Html/HtmlObject.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 // Requires: config.attributeName
 class HtmlObject extends HtmlRelation {
