+++ lib/relations/Html/HtmlOpenGraph.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlOpenGraph extends HtmlRelation {
   get href() {
