+++ lib/relations/Html/HtmlPictureSource.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlPictureSource extends HtmlRelation {
   get href() {
