+++ lib/relations/Html/HtmlPictureSourceSrcSet.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlPictureSourceSrcSet extends HtmlRelation {
   inline() {
