+++ lib/relations/Html/HtmlResourceHint.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 function getExtension(url) {
   const matches = url.match(/\.([^$#?]+)/);
