+++ lib/relations/Html/HtmlScript.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');;
 
 class HtmlScript extends HtmlRelation {
   get href() {
