+++ lib/relations/Html/HtmlSearchLink.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlSearchLink extends HtmlRelation {
   get href() {
