+++ lib/relations/Html/HtmlServiceWorkerRegistration.js
@@ -1,6 +1,6 @@
 // https://w3c.github.io/ServiceWorker/#link-type-serviceworker
 
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlServiceWorkerRegistration extends HtmlRelation {
   get href() {
