+++ lib/relations/Html/HtmlShortcutIcon.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlShortcutIcon extends HtmlRelation {
   get href() {
