+++ lib/relations/Html/HtmlStyle.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlStyle extends HtmlRelation {
   get href() {
