+++ lib/relations/Html/HtmlStyleAttribute.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlStyleAttribute extends HtmlRelation {
   get href() {
