+++ lib/relations/Html/HtmlSvgIsland.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlSvgIsland extends HtmlRelation {
   inline() {
