+++ lib/relations/Html/HtmlTemplate.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlTemplate extends HtmlRelation {
   inline() {
