+++ lib/relations/Html/HtmlVideo.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlVideo extends HtmlRelation {
   get href() {
