+++ lib/relations/Html/HtmlVideoPoster.js
@@ -1,4 +1,4 @@
-const HtmlRelation = require('./HtmlRelation');
+const HtmlRelation = require('../HtmlRelation');
 
 class HtmlVideoPoster extends HtmlRelation {
   get href() {
