+++ lib/relations/HtmlRelation.js
@@ -1,4 +1,4 @@
-const Relation = require('../Relation');
+const Relation = require('./Relation');
 
 /**
  * Base Relation for all types of HTML Relations
